warehouse/requirements.txt
django==2.0
djangorestframework
requests
pillow
gunicorn